More OS X networking changes.
git-svn-id: file:///home/jj/hercules.svn/trunk@2281 956126f8-22a0-4046-8f4a-272fa8102e63

+++ hercifc.h
@@ -43,14 +43,14 @@
 
 typedef struct _CTLREQ
 {
-    long   iType;
-    int    iProcID;
-    int    iCtlOp;
-    char   szIFName[IFNAMSIZ];
+    long                 iType;
+    int                  iProcID;
+    unsigned long int    iCtlOp;
+    char                 szIFName[IFNAMSIZ];
     union 
     {
-        struct ifreq   ifreq;
-        struct rtentry rtentry;
+        struct ifreq     ifreq;
+        struct rtentry   rtentry;
     } iru;
 }   CTLREQ, *PCTLREQ;
